feat: restrict SQLite API client access to the logged-in status only

+++ src/utils/sqliteChatIdApiClient.ts
@@ -2,67 +2,18 @@ import { Chat, ImageFile } from '@/src/types/chat';
 import { OptionType } from '@/src/types/common';
 
 /**
- * Create a new chat with the given title and metadata.
- * @param chatTitle - The title of the chat.
- * @param metadata - Additional metadata for the chat.
- * @returns The created chat or an error message if the creation failed.
- */
-export const updateChats = async (chatTitle: string, metadata: any) => {
-  try {
-    const res = await fetch('/api/chats', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        title: chatTitle,
-        metadata
-      })
-    });
-    if (!res) {
-      console.error('Error on post data to Chats table');
-    }
-    const chat = await res.json();
-
-    return chat;
-  } catch (e) {
-    console.error("Chat isn't created", e);
-  }
-};
-
-/**
- * Fetch all chats and return them as an array of OptionType.
- * @returns A Promise that resolves to an array of OptionType representing the chats.
- */
-export const fetchChats = async (): Promise<OptionType[]> => {
-  try {
-    const response = await fetch('/api/chats');
-    if (!response.ok) {
-      throw new Error(`Error: ${response.statusText}`);
-    }
-    const data = await response.json();
-    const chats = data.map((d: Chat, i: number) => ({
-      label: d.title,
-      value: d.id.toString()
-    }));
-    return chats;
-  } catch (error: any) {
-    console.error(`Failed to fetch chats: ${error.message}`);
-    return [];
-  }
-};
-
-/**
  * Deletes a chat with the given chatId.
  * @param chatId - The ID of the chat to delete.
  * @returns A promise that resolves to the result of the deletion.
  */
-export const deleteChat = async (chatId: string) => {
+export const deleteChat = async (token: string, chatId: string) => {
+  if (!token) return null;
   try {
     const res = await fetch(`/api/chats/${chatId}/chat`, {
       method: 'DELETE',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`
       }
     });
 
@@ -85,12 +36,18 @@ export const deleteChat = async (chatId: string) => {
  * @param newTitle - The new title for the chat.
  * @returns A promise that resolves to the result of the update.
  */
-export const editChatTitle = async (chatId: string, newTitle: string) => {
+export const editChatTitle = async (
+  token: string,
+  chatId: string,
+  newTitle: string
+) => {
+  if (!token) return null;
   try {
     const res = await fetch(`/api/chats/${chatId}/chat`, {
       method: 'PUT',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`
       },
       body: JSON.stringify({ title: newTitle })
     });
@@ -107,7 +64,6 @@ export const editChatTitle = async (chatId: string, newTitle: string) => {
     return null;
   }
 };
-
 /**
  * Create a new chat message with the given details.
  * @param userMessage - The message from the user.
@@ -116,28 +72,31 @@ export const editChatTitle = async (chatId: string, newTitle: string) => {
  * @returns The created chat message or an error message if the creation failed.
  */
 export const updateChatMessages = async (
+  token: string,
+  chatId: number,
   userMessage: string,
   aiMessage: string,
   model: string,
-  chatId: number,
   imageSrc: ImageFile[]
 ) => {
+  if (!token) return null;
   try {
     const res = await fetch(`/api/chats/${chatId}/messages`, {
       method: 'POST',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`
       },
       body: JSON.stringify({
         chatId,
         userMessage,
         aiMessage,
-        imageSrc,
-        model
+        model,
+        imageSrc
       })
     });
-    if (!res) {
-      console.error('Error on post message to ChatMessage table');
+    if (!res.ok) {
+      throw new Error(`Error: ${res.statusText}`);
     }
     const chatMessage = await res.json();
 
@@ -152,12 +111,15 @@ export const updateChatMessages = async (
  * @param chatId - The ID of the chat.
  * @returns A list of chat messages or an error message if the fetch failed.
  */
-export const fetchChatMessages = async (chatId: number) => {
+export const fetchChatMessages = async (token: string, chatId: number) => {
+  if (!token) return null;
+
   try {
     const res = await fetch(`/api/chats/${chatId}/messages`, {
       method: 'GET',
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`
       }
     });
